requirements.txt
# python==3.9.12
-i https://mirrors.aliyun.com/pypi/simple/
pandas==1.3.5
numpy==1.22.4
scipy==1.7.3
requests==2.20.0
fastparquet==0.8.3
matplotlib==3.8.4
joblib==1.2.0
loguru==0.7.2
jmespath
python-dotenv
deprecation
retrying==1.3.4
# zmail
